+++ packages/botonic-plugin-flow-builder/tests/tracking/track-knowledge-base.test.ts
@@ -29,7 +29,7 @@ describe('Check tracked events when a bot generates a response using a knowledge
           answer:
             'Flow Builder is a visual tool used to create and manage Conversational Apps. It allows users to design conversational flows by dragging and dropping elements, connecting them, and adding content to create conversational experiences. The tool is designed to enable non-technical users to create and manage Conversational Apps autonomously.',
           hasKnowledge: true,
-          isFaithuful: true,
+          isFaithful: true,
         }),
       },
       requestArgs: {